Testbench initial commit
Includes the the basic structure of the testbench, the
gui and all source files up to this date.
The testbench is able to charge and discharge a cell, to
measure the capacity, to record the open circuit voltage
curve, to charge the cell to a specific state of charge
and to run a given profile of different currents on the
cell.
The collected data is saved in a different folder '/data'
and includes the measured current, voltage and temperature
in correlation to the time.